+++ CHANGELOG.md
@@ -6,9 +6,12 @@ All notable changes to [bpmn-js-properties-panel](https://github.com/bpmn-io/bpm
 
 ___Note:__ Yet to be released changes appear here._
 
+## 1.16.0
+
 * `FEAT`: allow escalation code to be a FEEL expression ([#855](https://github.com/bpmn-io/bpmn-js-properties-panel/pull/855))
 * `FIX`: use FEEL editor for error code in throw events only
 * `FIX`: make it possible to clear `optional` `Dropdown` field defined through element template ([#834](https://github.com/bpmn-io/bpmn-js-properties-panel/issues/834))
+* `FIX`: remove task headers when implementation set to none ([#826](https://github.com/bpmn-io/bpmn-js-properties-panel/issues/826))
 
 ## 1.15.1